+++ crates/simulate/src/event_call.rs
@@ -1,5 +1,5 @@
+use data_decoder::DecodedValue;
 use serde::Serialize;
-use verification::CodeLocation;
 use walnut_shared::Parameter;
 
 #[derive(Debug, Serialize)]
@@ -8,12 +8,8 @@ pub struct EventCall {
     pub contract_call_id: u32,
 
     pub name: String,
-    pub keys: Vec<String>,
-    pub parameters: Vec<Parameter>,
-    pub data: Vec<String>,
-
-    pub debugger_trace_step_index: Option<usize>,
-    pub code_location: Option<CodeLocation>,
+    pub selector: String,
+    pub datas: Option<Vec<DecodedValue>>,
 
     pub is_hidden: bool,
 }